+++ consensus-types/light-client/header.go
@@ -4,11 +4,13 @@ import (
 	"fmt"
 
 	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
+	"github.com/prysmaticlabs/prysm/v5/config/params"
 	consensustypes "github.com/prysmaticlabs/prysm/v5/consensus-types"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
 	pb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/runtime/version"
+	"github.com/prysmaticlabs/prysm/v5/time/slots"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -22,6 +24,9 @@ func NewWrappedHeader(m proto.Message) (interfaces.LightClientHeader, error) {
 	case *pb.LightClientHeaderCapella:
 		return NewWrappedHeaderCapella(t)
 	case *pb.LightClientHeaderDeneb:
+		if slots.ToEpoch(t.Beacon.Slot) >= params.BeaconConfig().ElectraForkEpoch {
+			return NewWrappedHeaderElectra(t)
+		}
 		return NewWrappedHeaderDeneb(t)
 	default:
 		return nil, fmt.Errorf("cannot construct light client header from type %T", t)
@@ -53,6 +58,10 @@ func (h *headerAltair) SizeSSZ() int {
 	return h.p.SizeSSZ()
 }
 
+func (h *headerAltair) Proto() proto.Message {
+	return h.p
+}
+
 func (h *headerAltair) Version() int {
 	return version.Altair
 }
@@ -62,11 +71,11 @@ func (h *headerAltair) Beacon() *pb.BeaconBlockHeader {
 }
 
 func (h *headerAltair) Execution() (interfaces.ExecutionData, error) {
-	return nil, consensustypes.ErrNotSupported("Execution", version.Altair)
+	return nil, consensustypes.ErrNotSupported("Execution", h.Version())
 }
 
 func (h *headerAltair) ExecutionBranch() (interfaces.LightClientExecutionBranch, error) {
-	return interfaces.LightClientExecutionBranch{}, consensustypes.ErrNotSupported("ExecutionBranch", version.Altair)
+	return interfaces.LightClientExecutionBranch{}, consensustypes.ErrNotSupported("ExecutionBranch", h.Version())
 }
 
 type headerCapella struct {
@@ -112,6 +121,10 @@ func (h *headerCapella) MarshalSSZ() ([]byte, error) {
 
 func (h *headerCapella) SizeSSZ() int {
 	return h.p.SizeSSZ()
+}
+
+func (h *headerCapella) Proto() proto.Message {
+	return h.p
 }
 
 func (h *headerCapella) Version() int {
@@ -175,6 +188,10 @@ func (h *headerDeneb) SizeSSZ() int {
 	return h.p.SizeSSZ()
 }
 
+func (h *headerDeneb) Proto() proto.Message {
+	return h.p
+}
+
 func (h *headerDeneb) Version() int {
 	return version.Deneb
 }
@@ -188,5 +205,70 @@ func (h *headerDeneb) Execution() (interfaces.ExecutionData, error) {
 }
 
 func (h *headerDeneb) ExecutionBranch() (interfaces.LightClientExecutionBranch, error) {
+	return h.executionBranch, nil
+}
+
+type headerElectra struct {
+	p               *pb.LightClientHeaderDeneb
+	execution       interfaces.ExecutionData
+	executionBranch interfaces.LightClientExecutionBranch
+}
+
+var _ interfaces.LightClientHeader = &headerElectra{}
+
+func NewWrappedHeaderElectra(p *pb.LightClientHeaderDeneb) (interfaces.LightClientHeader, error) {
+	if p == nil {
+		return nil, consensustypes.ErrNilObjectWrapped
+	}
+	execution, err := blocks.WrappedExecutionPayloadHeaderDeneb(p.Execution)
+	if err != nil {
+		return nil, err
+	}
+
+	branch, err := createBranch[interfaces.LightClientExecutionBranch](
+		"execution",
+		p.ExecutionBranch,
+		fieldparams.ExecutionBranchDepth,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return &headerElectra{
+		p:               p,
+		execution:       execution,
+		executionBranch: branch,
+	}, nil
+}
+
+func (h *headerElectra) MarshalSSZTo(dst []byte) ([]byte, error) {
+	return h.p.MarshalSSZTo(dst)
+}
+
+func (h *headerElectra) MarshalSSZ() ([]byte, error) {
+	return h.p.MarshalSSZ()
+}
+
+func (h *headerElectra) SizeSSZ() int {
+	return h.p.SizeSSZ()
+}
+
+func (h *headerElectra) Proto() proto.Message {
+	return h.p
+}
+
+func (h *headerElectra) Version() int {
+	return version.Electra
+}
+
+func (h *headerElectra) Beacon() *pb.BeaconBlockHeader {
+	return h.p.Beacon
+}
+
+func (h *headerElectra) Execution() (interfaces.ExecutionData, error) {
+	return h.execution, nil
+}
+
+func (h *headerElectra) ExecutionBranch() (interfaces.LightClientExecutionBranch, error) {
 	return h.executionBranch, nil
 }