+++ consensus-types/light-client/update.go
@@ -23,10 +23,16 @@ func NewWrappedUpdate(m proto.Message) (interfaces.LightClientUpdate, error) {
 		return NewWrappedUpdateCapella(t)
 	case *pb.LightClientUpdateDeneb:
 		return NewWrappedUpdateDeneb(t)
+	case *pb.LightClientUpdateElectra:
+		return NewWrappedUpdateElectra(t)
 	default:
 		return nil, fmt.Errorf("cannot construct light client update from type %T", t)
 	}
 }
+
+// In addition to the proto object being wrapped, we store some fields that have to be
+// constructed from the proto, so that we don't have to reconstruct them every time
+// in getters.
 
 type updateAltair struct {
 	p                       *pb.LightClientUpdateAltair
@@ -42,14 +48,20 @@ func NewWrappedUpdateAltair(p *pb.LightClientUpdateAltair) (interfaces.LightClie
 	if p == nil {
 		return nil, consensustypes.ErrNilObjectWrapped
 	}
-	attestedHeader, err := NewWrappedHeaderAltair(p.AttestedHeader)
+
+	attestedHeader, err := NewWrappedHeader(p.AttestedHeader)
 	if err != nil {
 		return nil, err
 	}
-	finalizedHeader, err := NewWrappedHeaderAltair(p.FinalizedHeader)
-	if err != nil {
-		return nil, err
+
+	var finalizedHeader interfaces.LightClientHeader
+	if p.FinalizedHeader != nil {
+		finalizedHeader, err = NewWrappedHeader(p.FinalizedHeader)
+		if err != nil {
+			return nil, err
+		}
 	}
+
 	scBranch, err := createBranch[interfaces.LightClientSyncCommitteeBranch](
 		"sync committee",
 		p.NextSyncCommitteeBranch,
@@ -88,6 +100,10 @@ func (u *updateAltair) SizeSSZ() int {
 	return u.p.SizeSSZ()
 }
 
+func (u *updateAltair) Proto() proto.Message {
+	return u.p
+}
+
 func (u *updateAltair) Version() int {
 	return version.Altair
 }
@@ -96,12 +112,38 @@ func (u *updateAltair) AttestedHeader() interfaces.LightClientHeader {
 	return u.attestedHeader
 }
 
+func (u *updateAltair) SetAttestedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderAltair)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderAltair{})
+	}
+	u.p.AttestedHeader = proto
+	u.attestedHeader = header
+	return nil
+}
+
 func (u *updateAltair) NextSyncCommittee() *pb.SyncCommittee {
 	return u.p.NextSyncCommittee
 }
 
+func (u *updateAltair) SetNextSyncCommittee(sc *pb.SyncCommittee) {
+	u.p.NextSyncCommittee = sc
+}
+
 func (u *updateAltair) NextSyncCommitteeBranch() (interfaces.LightClientSyncCommitteeBranch, error) {
 	return u.nextSyncCommitteeBranch, nil
+}
+
+func (u *updateAltair) SetNextSyncCommitteeBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientSyncCommitteeBranch]("sync committee", branch, fieldparams.SyncCommitteeBranchDepth)
+	if err != nil {
+		return err
+	}
+	u.nextSyncCommitteeBranch = b
+
+	u.p.NextSyncCommitteeBranch = branch
+
+	return nil
 }
 
 func (u *updateAltair) NextSyncCommitteeBranchElectra() (interfaces.LightClientSyncCommitteeBranchElectra, error) {
@@ -112,18 +154,53 @@ func (u *updateAltair) FinalizedHeader() interfaces.LightClientHeader {
 	return u.finalizedHeader
 }
 
-func (u *updateAltair) FinalityBranch() interfaces.LightClientFinalityBranch {
-	return u.finalityBranch
+func (u *updateAltair) SetFinalizedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderAltair)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderAltair{})
+	}
+	u.p.FinalizedHeader = proto
+	u.finalizedHeader = header
+	return nil
+}
+
+func (u *updateAltair) FinalityBranch() (interfaces.LightClientFinalityBranch, error) {
+	return u.finalityBranch, nil
+}
+
+func (u *updateAltair) SetFinalityBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientFinalityBranch]("finality", branch, fieldparams.FinalityBranchDepth)
+	if err != nil {
+		return err
+	}
+	u.finalityBranch = b
+	u.p.FinalityBranch = branch
+	return nil
+}
+
+func (u *updateAltair) FinalityBranchElectra() (interfaces.LightClientFinalityBranchElectra, error) {
+	return interfaces.LightClientFinalityBranchElectra{}, consensustypes.ErrNotSupported("FinalityBranchElectra", version.Altair)
 }
 
 func (u *updateAltair) SyncAggregate() *pb.SyncAggregate {
 	return u.p.SyncAggregate
 }
 
+func (u *updateAltair) SetSyncAggregate(sa *pb.SyncAggregate) {
+	u.p.SyncAggregate = sa
+}
+
 func (u *updateAltair) SignatureSlot() primitives.Slot {
 	return u.p.SignatureSlot
 }
 
+func (u *updateAltair) SetSignatureSlot(slot primitives.Slot) {
+	u.p.SignatureSlot = slot
+}
+
+// In addition to the proto object being wrapped, we store some fields that have to be
+// constructed from the proto, so that we don't have to reconstruct them every time
+// in getters.
 type updateCapella struct {
 	p                       *pb.LightClientUpdateCapella
 	attestedHeader          interfaces.LightClientHeader
@@ -138,14 +215,20 @@ func NewWrappedUpdateCapella(p *pb.LightClientUpdateCapella) (interfaces.LightCl
 	if p == nil {
 		return nil, consensustypes.ErrNilObjectWrapped
 	}
-	attestedHeader, err := NewWrappedHeaderCapella(p.AttestedHeader)
+
+	attestedHeader, err := NewWrappedHeader(p.AttestedHeader)
 	if err != nil {
 		return nil, err
 	}
-	finalizedHeader, err := NewWrappedHeaderCapella(p.FinalizedHeader)
-	if err != nil {
-		return nil, err
+
+	var finalizedHeader interfaces.LightClientHeader
+	if p.FinalizedHeader != nil {
+		finalizedHeader, err = NewWrappedHeader(p.FinalizedHeader)
+		if err != nil {
+			return nil, err
+		}
 	}
+
 	scBranch, err := createBranch[interfaces.LightClientSyncCommitteeBranch](
 		"sync committee",
 		p.NextSyncCommitteeBranch,
@@ -184,6 +267,10 @@ func (u *updateCapella) SizeSSZ() int {
 	return u.p.SizeSSZ()
 }
 
+func (u *updateCapella) Proto() proto.Message {
+	return u.p
+}
+
 func (u *updateCapella) Version() int {
 	return version.Capella
 }
@@ -192,12 +279,38 @@ func (u *updateCapella) AttestedHeader() interfaces.LightClientHeader {
 	return u.attestedHeader
 }
 
+func (u *updateCapella) SetAttestedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderCapella)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderCapella{})
+	}
+	u.p.AttestedHeader = proto
+	u.attestedHeader = header
+	return nil
+}
+
 func (u *updateCapella) NextSyncCommittee() *pb.SyncCommittee {
 	return u.p.NextSyncCommittee
 }
 
+func (u *updateCapella) SetNextSyncCommittee(sc *pb.SyncCommittee) {
+	u.p.NextSyncCommittee = sc
+}
+
 func (u *updateCapella) NextSyncCommitteeBranch() (interfaces.LightClientSyncCommitteeBranch, error) {
 	return u.nextSyncCommitteeBranch, nil
+}
+
+func (u *updateCapella) SetNextSyncCommitteeBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientSyncCommitteeBranch]("sync committee", branch, fieldparams.SyncCommitteeBranchDepth)
+	if err != nil {
+		return err
+	}
+	u.nextSyncCommitteeBranch = b
+
+	u.p.NextSyncCommitteeBranch = branch
+
+	return nil
 }
 
 func (u *updateCapella) NextSyncCommitteeBranchElectra() (interfaces.LightClientSyncCommitteeBranchElectra, error) {
@@ -208,18 +321,53 @@ func (u *updateCapella) FinalizedHeader() interfaces.LightClientHeader {
 	return u.finalizedHeader
 }
 
-func (u *updateCapella) FinalityBranch() interfaces.LightClientFinalityBranch {
-	return u.finalityBranch
+func (u *updateCapella) SetFinalizedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderCapella)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderCapella{})
+	}
+	u.p.FinalizedHeader = proto
+	u.finalizedHeader = header
+	return nil
+}
+
+func (u *updateCapella) FinalityBranch() (interfaces.LightClientFinalityBranch, error) {
+	return u.finalityBranch, nil
+}
+
+func (u *updateCapella) SetFinalityBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientFinalityBranch]("finality", branch, fieldparams.FinalityBranchDepth)
+	if err != nil {
+		return err
+	}
+	u.finalityBranch = b
+	u.p.FinalityBranch = branch
+	return nil
+}
+
+func (u *updateCapella) FinalityBranchElectra() (interfaces.LightClientFinalityBranchElectra, error) {
+	return interfaces.LightClientFinalityBranchElectra{}, consensustypes.ErrNotSupported("FinalityBranchElectra", u.Version())
 }
 
 func (u *updateCapella) SyncAggregate() *pb.SyncAggregate {
 	return u.p.SyncAggregate
 }
 
+func (u *updateCapella) SetSyncAggregate(sa *pb.SyncAggregate) {
+	u.p.SyncAggregate = sa
+}
+
 func (u *updateCapella) SignatureSlot() primitives.Slot {
 	return u.p.SignatureSlot
 }
 
+func (u *updateCapella) SetSignatureSlot(slot primitives.Slot) {
+	u.p.SignatureSlot = slot
+}
+
+// In addition to the proto object being wrapped, we store some fields that have to be
+// constructed from the proto, so that we don't have to reconstruct them every time
+// in getters.
 type updateDeneb struct {
 	p                       *pb.LightClientUpdateDeneb
 	attestedHeader          interfaces.LightClientHeader
@@ -234,14 +382,20 @@ func NewWrappedUpdateDeneb(p *pb.LightClientUpdateDeneb) (interfaces.LightClient
 	if p == nil {
 		return nil, consensustypes.ErrNilObjectWrapped
 	}
-	attestedHeader, err := NewWrappedHeaderDeneb(p.AttestedHeader)
+
+	attestedHeader, err := NewWrappedHeader(p.AttestedHeader)
 	if err != nil {
 		return nil, err
 	}
-	finalizedHeader, err := NewWrappedHeaderDeneb(p.FinalizedHeader)
-	if err != nil {
-		return nil, err
+
+	var finalizedHeader interfaces.LightClientHeader
+	if p.FinalizedHeader != nil {
+		finalizedHeader, err = NewWrappedHeader(p.FinalizedHeader)
+		if err != nil {
+			return nil, err
+		}
 	}
+
 	scBranch, err := createBranch[interfaces.LightClientSyncCommitteeBranch](
 		"sync committee",
 		p.NextSyncCommitteeBranch,
@@ -280,6 +434,10 @@ func (u *updateDeneb) SizeSSZ() int {
 	return u.p.SizeSSZ()
 }
 
+func (u *updateDeneb) Proto() proto.Message {
+	return u.p
+}
+
 func (u *updateDeneb) Version() int {
 	return version.Deneb
 }
@@ -288,12 +446,38 @@ func (u *updateDeneb) AttestedHeader() interfaces.LightClientHeader {
 	return u.attestedHeader
 }
 
+func (u *updateDeneb) SetAttestedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderDeneb)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderDeneb{})
+	}
+	u.p.AttestedHeader = proto
+	u.attestedHeader = header
+	return nil
+}
+
 func (u *updateDeneb) NextSyncCommittee() *pb.SyncCommittee {
 	return u.p.NextSyncCommittee
 }
 
+func (u *updateDeneb) SetNextSyncCommittee(sc *pb.SyncCommittee) {
+	u.p.NextSyncCommittee = sc
+}
+
 func (u *updateDeneb) NextSyncCommitteeBranch() (interfaces.LightClientSyncCommitteeBranch, error) {
 	return u.nextSyncCommitteeBranch, nil
+}
+
+func (u *updateDeneb) SetNextSyncCommitteeBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientSyncCommitteeBranch]("sync committee", branch, fieldparams.SyncCommitteeBranchDepth)
+	if err != nil {
+		return err
+	}
+	u.nextSyncCommitteeBranch = b
+
+	u.p.NextSyncCommitteeBranch = branch
+
+	return nil
 }
 
 func (u *updateDeneb) NextSyncCommitteeBranchElectra() (interfaces.LightClientSyncCommitteeBranchElectra, error) {
@@ -304,24 +488,59 @@ func (u *updateDeneb) FinalizedHeader() interfaces.LightClientHeader {
 	return u.finalizedHeader
 }
 
-func (u *updateDeneb) FinalityBranch() interfaces.LightClientFinalityBranch {
-	return u.finalityBranch
+func (u *updateDeneb) SetFinalizedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderDeneb)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderDeneb{})
+	}
+	u.p.FinalizedHeader = proto
+	u.finalizedHeader = header
+	return nil
+}
+
+func (u *updateDeneb) FinalityBranch() (interfaces.LightClientFinalityBranch, error) {
+	return u.finalityBranch, nil
+}
+
+func (u *updateDeneb) SetFinalityBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientFinalityBranch]("finality", branch, fieldparams.FinalityBranchDepth)
+	if err != nil {
+		return err
+	}
+	u.finalityBranch = b
+	u.p.FinalityBranch = branch
+	return nil
+}
+
+func (u *updateDeneb) FinalityBranchElectra() (interfaces.LightClientFinalityBranchElectra, error) {
+	return interfaces.LightClientFinalityBranchElectra{}, consensustypes.ErrNotSupported("FinalityBranchElectra", u.Version())
 }
 
 func (u *updateDeneb) SyncAggregate() *pb.SyncAggregate {
 	return u.p.SyncAggregate
 }
 
+func (u *updateDeneb) SetSyncAggregate(sa *pb.SyncAggregate) {
+	u.p.SyncAggregate = sa
+}
+
 func (u *updateDeneb) SignatureSlot() primitives.Slot {
 	return u.p.SignatureSlot
 }
 
+func (u *updateDeneb) SetSignatureSlot(slot primitives.Slot) {
+	u.p.SignatureSlot = slot
+}
+
+// In addition to the proto object being wrapped, we store some fields that have to be
+// constructed from the proto, so that we don't have to reconstruct them every time
+// in getters.
 type updateElectra struct {
 	p                       *pb.LightClientUpdateElectra
 	attestedHeader          interfaces.LightClientHeader
 	nextSyncCommitteeBranch interfaces.LightClientSyncCommitteeBranchElectra
 	finalizedHeader         interfaces.LightClientHeader
-	finalityBranch          interfaces.LightClientFinalityBranch
+	finalityBranch          interfaces.LightClientFinalityBranchElectra
 }
 
 var _ interfaces.LightClientUpdate = &updateElectra{}
@@ -330,14 +549,20 @@ func NewWrappedUpdateElectra(p *pb.LightClientUpdateElectra) (interfaces.LightCl
 	if p == nil {
 		return nil, consensustypes.ErrNilObjectWrapped
 	}
-	attestedHeader, err := NewWrappedHeaderDeneb(p.AttestedHeader)
+
+	attestedHeader, err := NewWrappedHeader(p.AttestedHeader)
 	if err != nil {
 		return nil, err
 	}
-	finalizedHeader, err := NewWrappedHeaderDeneb(p.FinalizedHeader)
-	if err != nil {
-		return nil, err
+
+	var finalizedHeader interfaces.LightClientHeader
+	if p.FinalizedHeader != nil {
+		finalizedHeader, err = NewWrappedHeader(p.FinalizedHeader)
+		if err != nil {
+			return nil, err
+		}
 	}
+
 	scBranch, err := createBranch[interfaces.LightClientSyncCommitteeBranchElectra](
 		"sync committee",
 		p.NextSyncCommitteeBranch,
@@ -346,10 +571,11 @@ func NewWrappedUpdateElectra(p *pb.LightClientUpdateElectra) (interfaces.LightCl
 	if err != nil {
 		return nil, err
 	}
-	finalityBranch, err := createBranch[interfaces.LightClientFinalityBranch](
+
+	finalityBranch, err := createBranch[interfaces.LightClientFinalityBranchElectra](
 		"finality",
 		p.FinalityBranch,
-		fieldparams.FinalityBranchDepth,
+		fieldparams.FinalityBranchDepthElectra,
 	)
 	if err != nil {
 		return nil, err
@@ -376,6 +602,10 @@ func (u *updateElectra) SizeSSZ() int {
 	return u.p.SizeSSZ()
 }
 
+func (u *updateElectra) Proto() proto.Message {
+	return u.p
+}
+
 func (u *updateElectra) Version() int {
 	return version.Electra
 }
@@ -384,12 +614,38 @@ func (u *updateElectra) AttestedHeader() interfaces.LightClientHeader {
 	return u.attestedHeader
 }
 
+func (u *updateElectra) SetAttestedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderDeneb)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderDeneb{})
+	}
+	u.p.AttestedHeader = proto
+	u.attestedHeader = header
+	return nil
+}
+
 func (u *updateElectra) NextSyncCommittee() *pb.SyncCommittee {
 	return u.p.NextSyncCommittee
 }
 
+func (u *updateElectra) SetNextSyncCommittee(sc *pb.SyncCommittee) {
+	u.p.NextSyncCommittee = sc
+}
+
 func (u *updateElectra) NextSyncCommitteeBranch() (interfaces.LightClientSyncCommitteeBranch, error) {
 	return [5][32]byte{}, consensustypes.ErrNotSupported("NextSyncCommitteeBranch", version.Electra)
+}
+
+func (u *updateElectra) SetNextSyncCommitteeBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientSyncCommitteeBranchElectra]("sync committee", branch, fieldparams.SyncCommitteeBranchDepthElectra)
+	if err != nil {
+		return err
+	}
+	u.nextSyncCommitteeBranch = b
+
+	u.p.NextSyncCommitteeBranch = branch
+
+	return nil
 }
 
 func (u *updateElectra) NextSyncCommitteeBranchElectra() (interfaces.LightClientSyncCommitteeBranchElectra, error) {
@@ -400,14 +656,46 @@ func (u *updateElectra) FinalizedHeader() interfaces.LightClientHeader {
 	return u.finalizedHeader
 }
 
-func (u *updateElectra) FinalityBranch() interfaces.LightClientFinalityBranch {
-	return u.finalityBranch
+func (u *updateElectra) SetFinalizedHeader(header interfaces.LightClientHeader) error {
+	proto, ok := header.Proto().(*pb.LightClientHeaderDeneb)
+	if !ok {
+		return fmt.Errorf("header type %T is not %T", proto, &pb.LightClientHeaderDeneb{})
+	}
+	u.p.FinalizedHeader = proto
+	u.finalizedHeader = header
+	return nil
+}
+
+func (u *updateElectra) FinalityBranch() (interfaces.LightClientFinalityBranch, error) {
+	return interfaces.LightClientFinalityBranch{}, consensustypes.ErrNotSupported("FinalityBranch", u.Version())
+}
+
+func (u *updateElectra) SetFinalityBranch(branch [][]byte) error {
+	b, err := createBranch[interfaces.LightClientFinalityBranchElectra]("finality", branch, fieldparams.FinalityBranchDepthElectra)
+	if err != nil {
+		return err
+	}
+	u.finalityBranch = b
+	u.p.FinalityBranch = branch
+	return nil
+}
+
+func (u *updateElectra) FinalityBranchElectra() (interfaces.LightClientFinalityBranchElectra, error) {
+	return u.finalityBranch, nil
 }
 
 func (u *updateElectra) SyncAggregate() *pb.SyncAggregate {
 	return u.p.SyncAggregate
 }
 
+func (u *updateElectra) SetSyncAggregate(sa *pb.SyncAggregate) {
+	u.p.SyncAggregate = sa
+}
+
 func (u *updateElectra) SignatureSlot() primitives.Slot {
 	return u.p.SignatureSlot
+}
+
+func (u *updateElectra) SetSignatureSlot(slot primitives.Slot) {
+	u.p.SignatureSlot = slot
 }